Minor updates
Former-commit-id: c52db0854056b13e9cb5d7fc30a0ffb10f7e755e
Former-commit-id: 50164183fe73a3371fc5beab3a23a3f82fcf4181

+++ src/utils/optimus_prime.py
@@ -136,32 +136,33 @@ class Prime(commands.Cog):
         '''
         Catch people trying to edit messages
         '''
-        # moderation system
-        ctx = await self.bot.get_context(message)
+        if not message.author.bot and message.guild:
+            # moderation system
+            ctx = await self.bot.get_context(message)
 
-        if self.filter(ctx, message.content) or await self.additionalChecks(message):
-            await message.delete()
-        else:
-            try: self.bucket[message.guild.id][message.author.id] += [message]
-            except:
-                try:
-                    self.bucket[message.guild.id] = {}
-                    self.bucket[message.guild.id][message.author.id] = [message]
+            if self.filter(ctx, message.content) or await self.additionalChecks(message):
+                await message.delete()
+            else:
+                try: self.bucket[message.guild.id][message.author.id] += [message]
                 except:
-                    pass
-            
-            joined = "".join([m.content for m in self.bucket[message.guild.id][message.author.id]])
-
-            if self.filter(ctx, joined):
-                for m in self.bucket[message.guild.id][message.author.id]:
                     try:
-                        await m.delete()
+                        self.bucket[message.guild.id] = {}
+                        self.bucket[message.guild.id][message.author.id] = [message]
                     except:
-                        print("Cannot delete message {}".format(m.content))
-                self.bucket[message.guild.id][message.author.id] = []
-            
-            elif len(self.bucket[message.guild.id][message.author.id]) > MSG_BUFFER_LIMIT:
-                self.bucket[message.guild.id][message.author.id].pop(0)
+                        pass
+                
+                joined = "".join([m.content for m in self.bucket[message.guild.id][message.author.id]])
+
+                if self.filter(ctx, joined):
+                    for m in self.bucket[message.guild.id][message.author.id]:
+                        try:
+                            await m.delete()
+                        except:
+                            print("Cannot delete message {}".format(m.content))
+                    self.bucket[message.guild.id][message.author.id] = []
+                
+                elif len(self.bucket[message.guild.id][message.author.id]) > MSG_BUFFER_LIMIT:
+                    self.bucket[message.guild.id][message.author.id].pop(0)
 
 
     @commands.Cog.listener()
